pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-salesforce-architect"
version = "0.1.0"
description = "RAG system for Salesforce architecture & dev documentation (grounded answers with citations)."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "faiss-cpu>=1.13.2",
  "numpy>=1.25.0",
  "pydantic>=2.0.0",
  "python-dotenv>=1.0.0",
  "fastapi>=0.110.0",
  "uvicorn>=0.27.0",
  "pypdf>=4.0.0",
  "markdown>=3.6",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.6.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]